Persist synced block number (#173)
* Add synced block to ChainState

* Add Storage SyncedBlock functions

* Set SyncedBlock on SetChainState

* Add pagination for sync registered accounts

* Add endBlock to SyncBatches functions

* Add SyncSize to config

* Revert removing ChainConnector GetLatestBlockNumber func

* Fix ethereum config

* Simplify SyncOnStart function

* Use syncedBlock for chain queries

* Remove FilterPubkeyRegistered from WatchAccounts

* Move InitialSync function before workers

* Move InitialSync to separate file

* Improve InitialSync test

* Check batch sync in InitialSync test

* Process initial sync in pararell

* Extract calculation of end block

* Change channel to buffered

* Remove InitSync and move syncing to newBlockLoop

* Add start block to SyncBatches function

* Move rollupLoop to newBlockLoop

* Remove Storage proposer related functions

* Remove InitialSyncTests

* Remove WatchAccounts commander function

* Extract newBlockIteration

* Add TODOs

* Return uin64 instead of uint32 in GetLatestBlockNumber

* Sync to latestBlockNumber before subscribing to new blocks

* Pass context to stop rollup if needed

* Remove unused dependency and make stylistic changes

* Fix E2E test race condition by refactoring newBlockLoop

* Protect commander state critical section with mutex

* Make ganache mine block every second, decrease e2e bench txs count

* Improve a console log

* Increase docker commander start timeout

* Fix lint

+++ api/calculate_transfer_status_test.go
@@ -115,7 +115,7 @@ func (s *CalculateTransferStatusTestSuite) TestCalculateTransferStatus_Finalised
 	currentBlockNumber, err := s.sim.GetLatestBlockNumber()
 	s.NoError(err)
 	batch := models.Batch{
-		FinalisationBlock: ref.Uint32(*currentBlockNumber + 1),
+		FinalisationBlock: ref.Uint32(uint32(*currentBlockNumber) + 1),
 	}
 	batchID, err := s.storage.AddBatch(&batch)
 	s.NoError(err)
@@ -131,7 +131,7 @@ func (s *CalculateTransferStatusTestSuite) TestCalculateTransferStatus_Finalised
 	latestBlockNumber, err := s.sim.GetLatestBlockNumber()
 	s.NoError(err)
 
-	status, err := CalculateTransferStatus(s.storage, &s.transfer.TransactionBase, *latestBlockNumber)
+	status, err := CalculateTransferStatus(s.storage, &s.transfer.TransactionBase, uint32(*latestBlockNumber))
 	s.NoError(err)
 
 	s.Equal(txstatus.Finalised, *status)
